* master: (126 commits)
  update readme and changelog
  Update changelogs
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/16b4d4bc3f4da8825e7c4cc5b6e2627f21aee383
  Replace TBD entries with version 5.24.1
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/bf85214053f6e4356c3ba25f35cc83502cbdba7e
  Process changelog for refs/heads/release/T25.falcon This is an automated PR created by Camwyn. Generated by: https://github.com/the-events-calendar/event-tickets/actions/runs/15570142060
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/d568c5980c99e45e4b42c634c3719f282dbb5897
  :symbols: Updating .pot file and add language changelog for de0bbc87a1d3cc17edf3cbeb40bc3e35cbac95ca
  Stop using deprecated function tribe_required
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/89cbf971b73c7ccd9d4d9934c071ff5ef570453d
  Fix is_classic_editor in Overwrite
  update snapshots
  Undo earlier fix for ticketForm
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/77757abdcdc18103b0eb37d4757abec32166dc5f
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/caf4e55659ec0ec8d8f47be664766b500f93c8c3
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/d3e3e7d575a2e877f277d7878b8cf54535bbd2dc
  phpcs
  phpcs
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/d270444e17c724cfcba96a7331c0f75262f238b2
  :fast_forward: https://github.com/the-events-calendar/tribe-common/commit/f9cc2316ff379acf2964cf523d81df727790fb9b
  ...

Resolved Conflicts:
	event-tickets.php
	package.json
	readme.txt
	src/Tribe/Editor/Template/Overwrite.php
	src/Tribe/Main.php

+++ src/Tickets/Commerce/Ticket.php
@@ -2,6 +2,7 @@
 
 namespace TEC\Tickets\Commerce;
 
+use WP_Post;
 use TEC\Tickets\Commerce\Status\Denied;
 use TEC\Tickets\Commerce\Status\Pending;
 use TEC\Tickets\Commerce\Status\Status_Handler;
@@ -262,7 +263,7 @@ class Ticket extends Ticket_Data {
 	 *
 	 * @param Status_Interface      $new_status New post status.
 	 * @param Status_Interface|null $old_status Old post status.
-	 * @param \WP_Post              $post       Post object.
+	 * @param WP_Post               $post       Post object.
 	 */
 	public function modify_counters_by_status( $new_status, $old_status, $post ) {
 		$order = tec_tc_get_order( $post );
@@ -533,9 +534,9 @@ class Ticket extends Ticket_Data {
 	 *
 	 * @since 5.1.9
 	 *
-	 * @param       $post_id
-	 * @param       $ticket
-	 * @param array $raw_data
+	 * @param int|WP_Post|null $post_id   The post ID.
+	 * @param Ticket_Object    $ticket    The ticket object.
+	 * @param array            $raw_data  The raw data.
 	 *
 	 * @return false|int|\WP_Error
 	 */
@@ -552,7 +553,7 @@ class Ticket extends Ticket_Data {
 				'post_author'  => get_current_user_id(),
 				'post_excerpt' => $ticket->description,
 				'post_title'   => $ticket->name,
-				'menu_order'   => $ticket->menu_order ?? tribe_get_request_var( 'menu_order', - 1 ),
+				'menu_order'   => (int) ( $ticket->menu_order ?? tribe_get_request_var( 'menu_order', - 1 ) ),
 				'meta_input' => [
 					'_type' => $raw_data['ticket_type'] ?? 'default',
 				]
@@ -825,8 +826,8 @@ class Ticket extends Ticket_Data {
 	 *
 	 * @since 5.1.9
 	 *
-	 * @param $event_id
-	 * @param $ticket_id
+	 * @param int|WP_Post|null $event_id   The event ID.
+	 * @param int|WP_Post|null $ticket_id  The ticket ID.
 	 *
 	 * @return bool
 	 */
@@ -991,14 +992,14 @@ class Ticket extends Ticket_Data {
 	/**
 	 * Gets the product price value object
 	 *
-	 * @since   5.1.9
-	 * @since   5.2.3 method signature changed to return an instance of Value instead of a string.
-	 * @since   5.13.0   added new param to force regular price value return.
+	 * @since 5.1.9
+	 * @since 5.2.3 method signature changed to return an instance of Value instead of a string.
+	 * @since 5.13.0   added new param to force regular price value return.
 	 *
-	 * @param int|\WP_Post $product       The ticket post ID or object.
-	 * @param bool         $force_regular Whether to force the regular price.
+	 * @param int|WP_Post|null $product       The ticket post ID or object.
+	 * @param bool             $force_regular Whether to force the regular price.
 	 *
-	 * @return Commerce\Utils\Value;
+	 * @return \TEC\Tickets\Commerce\Utils\Value|null;
 	 * @version 5.2.3
 	 *
 	 */
@@ -1099,8 +1100,8 @@ class Ticket extends Ticket_Data {
 	 *
 	 * @since 5.5.10
 	 *
-	 * @param $ticket_id int The ticket post ID.
-	 * @param $quantity  int The quantity to increase the ticket stock by.
+	 * @param int|WP_Post|null $ticket_id  The ticket post ID.
+	 * @param int              $quantity   The quantity to increase the ticket stock by.
 	 *
 	 * @return bool|int
 	 */
